Rename services to kebab-case with type extension
