+++ src/controllers/company-add.controller.js
@@ -3,9 +3,9 @@ const Q = require('q')
 const winston = require('winston')
 const { ukOtherCompanyOptions, foreignOtherCompanyOptions } = require('../options')
 const { isBlank, toQueryString, genCSRF } = require('../lib/controllerutils')
-const searchService = require('../services/searchservice')
-const companyService = require('../services/companyservice')
-const companyFormattingService = require('../services/companyformattingservice')
+const searchService = require('../services/search.service')
+const companyService = require('../services/company.service')
+const companyFormattingService = require('../services/company-formatting.service')
 const { companyDetailsLabels, chDetailsLabels, companyTypeOptions } = require('../labels/companylabels')
 
 const router = express.Router()
