+++ src/controllers/company-archive.controller.js
@@ -3,7 +3,7 @@ const express = require('express')
 const winston = require('winston')
 const Q = require('q')
 const companyRepository = require('../repositorys/companyrepository')
-const companyService = require('../services/companyservice')
+const companyService = require('../services/company.service')
 const router = express.Router()
 
 function archiveCompany (req, res, next) {
