+++ src/controllers/company-foreign.controller.js
@@ -3,9 +3,9 @@ const express = require('express')
 const winston = require('winston')
 const Q = require('q')
 const companyRepository = require('../repositorys/companyrepository')
-const companyFormService = require('../services/companyformservice')
-const companyService = require('../services/companyservice')
-const companyFormattingService = require('../services/companyformattingservice')
+const companyFormService = require('../services/company-form.service')
+const companyService = require('../services/company.service')
+const companyFormattingService = require('../services/company-formatting.service')
 const { companyDetailsLabels, accountManagementDisplayLabels, hqLabels } = require('../labels/companylabels')
 const metadataRepository = require('../repositorys/metadatarepository')
 const { genCSRF, containsFormData, isBlank } = require('../lib/controllerutils')
