+++ src/controllers/company-investment-summary.controller.js
@@ -3,7 +3,7 @@ const express = require('express')
 const { isBlank } = require('../lib/controllerutils')
 const companyRepository = require('../repositorys/companyrepository')
 const metadataRepository = require('../repositorys/metadatarepository')
-const investmentFormattingService = require('../services/investmentformattingservice')
+const investmentFormattingService = require('../services/investment-formatting.service')
 const { investmentDetailLabels, investmentProjectsOpenLabels, investmentProjectsClosedLabels, investmentFormLabels } = require('../labels/investmentlabels')
 const { managedOptions, investmentTierOptions } = require('../options')
 const router = express.Router()
