+++ src/controllers/company-ltd.controller.js
@@ -2,12 +2,12 @@
 const express = require('express')
 const winston = require('winston')
 const Q = require('q')
-const companyService = require('../services/companyservice')
+const companyService = require('../services/company.service')
+const companyFormService = require('../services/company-form.service')
+const companyFormattingService = require('../services/company-formatting.service')
 const companyRepository = require('../repositorys/companyrepository')
 const metadataRepository = require('../repositorys/metadatarepository')
-const companyFormattingService = require('../services/companyformattingservice')
 const { companyDetailsLabels, chDetailsLabels, accountManagementDisplayLabels, hqLabels } = require('../labels/companylabels')
-const companyFormService = require('../services/companyformservice')
 const { genCSRF, containsFormData, isBlank } = require('../lib/controllerutils')
 const router = express.Router()
 const companyWithCHKeys = ['alias', 'trading_address', 'uk_region', 'headquarter_type', 'sector', 'website', 'description', 'employee_range', 'turnover_range']
