+++ src/controllers/contact-edit.controller.js
@@ -2,7 +2,7 @@ const express = require('express')
 const Q = require('q')
 const winston = require('winston')
 const { genCSRF, containsFormData } = require('../lib/controllerutils')
-const contactFormService = require('../services/contactformservice')
+const contactFormService = require('../services/contact-form.service')
 const { contactLabels } = require('../labels/contactlabels')
 const metadataRepository = require('../repositorys/metadatarepository')
 const companyRepository = require('../repositorys/companyrepository')
