+++ src/controllers/interaction-edit.controller.js
@@ -6,8 +6,8 @@ const contactRepository = require('../repositorys/contactrepository')
 const companyRepository = require('../repositorys/companyrepository')
 const metadataRepository = require('../repositorys/metadatarepository')
 const advisorRepository = require('../repositorys/advisorrepository')
-const interactionDataService = require('../services/interactiondataservice')
-const interactionFormService = require('../services/interactionformservice')
+const interactionDataService = require('../services/interaction-data.service')
+const interactionFormService = require('../services/interaction-form.service')
 const { genCSRF, containsFormData } = require('../lib/controllerutils')
 const router = express.Router()
