+++ src/controllers/service-delivery.controller.js
@@ -7,9 +7,9 @@ const { genCSRF, transformV2Errors } = require('../lib/controllerutils')
 const { nullEmptyFields, deleteNulls } = require('../lib/propertyhelpers')
 const metadataRepository = require('../repositorys/metadatarepository')
 const serviceDeliveryRepository = require('../repositorys/servicedeliveryrepository')
-const serviceDeliveryService = require('../services/servicedeliveryservice')
-const { getDisplayServiceDelivery } = require('../services/servicedeliveryformattingservice')
-const { getViewCompanyLink } = require('../services/companyservice')
+const serviceDeliveryService = require('../services/service-delivery.service')
+const { getDisplayServiceDelivery } = require('../services/service-delivery-formatting.service')
+const { getViewCompanyLink } = require('../services/company.service')
 
 const serviceDeliveryDisplayOrder = ['company', 'dit_team', 'service', 'status', 'subject', 'notes', 'date', 'dit_advisor', 'uk_region', 'sector', 'contact', 'country_of_interest']
 const router = express.Router()
