+++ test/services/interaction-data.service.test.js
@@ -25,7 +25,7 @@ describe('interaction data service', function () {
     getDitCompanyStub = sinon.stub().resolves(company)
     getInteractionStub = sinon.stub().resolves(interaction)
 
-    interactionDataService = proxyquire('../../src/services/interactiondataservice', {
+    interactionDataService = proxyquire('../../src/services/interaction-data.service', {
       '../repositorys/companyrepository': {
         getDitCompany: getDitCompanyStub
       },
@@ -54,7 +54,7 @@ describe('interaction data service', function () {
     it('should throw an error if fetching something fails', function (done) {
       getInteractionStub = sinon.stub().rejects(new Error('error'))
 
-      interactionDataService = proxyquire('../../src/services/interactiondataservice', {
+      interactionDataService = proxyquire('../../src/services/interaction-data.service', {
         '../repositorys/companyrepository': {
           getDitCompany: getDitCompanyStub
         },
@@ -110,7 +110,7 @@ describe('interaction data service', function () {
     it('should throw an error if something goes wrong', function (done) {
       getContactStub = sinon.stub().rejects(new Error('error'))
 
-      interactionDataService = proxyquire('../../src/services/interactiondataservice', {
+      interactionDataService = proxyquire('../../src/services/interaction-data.service', {
         '../repositorys/companyrepository': {
           getDitCompany: getDitCompanyStub
         },
@@ -133,7 +133,7 @@ describe('interaction data service', function () {
     it('should throw null for a contact with an invalid company', function (done) {
       getContactStub = sinon.stub().rejects(new Error('error'))
 
-      interactionDataService = proxyquire('../../src/services/interactiondataservice', {
+      interactionDataService = proxyquire('../../src/services/interaction-data.service', {
         '../repositorys/companyrepository': {
           getDitCompany: getDitCompanyStub
         },
@@ -178,7 +178,7 @@ describe('interaction data service', function () {
     it('should throw an error if something goes wrong', function (done) {
       getDitCompanyStub = sinon.stub().rejects(new Error('error'))
 
-      interactionDataService = proxyquire('../../src/services/interactiondataservice', {
+      interactionDataService = proxyquire('../../src/services/interaction-data.service', {
         '../repositorys/companyrepository': {
           getDitCompany: getDitCompanyStub
         },
