+++ test/services/interaction-form.service.test.js
@@ -36,7 +36,7 @@ describe('interaction form service', function () {
 
     saveInteractionStub = sinon.stub().resolves({ id: '1234', subject: 'subject', company: company.id, contact: contact.id })
 
-    interactionFormService = proxyquire('../../src/services/interactionformservice', {
+    interactionFormService = proxyquire('../../src/services/interaction-form.service', {
       '../repositorys/interactionrepository': {
         saveInteraction: saveInteractionStub
       }
@@ -131,7 +131,7 @@ describe('interaction form service', function () {
     it('should pass back any failures', function (done) {
       saveInteractionStub = sinon.stub().rejects(new Error('error'))
 
-      interactionFormService = proxyquire('../../src/services/interactionformservice', {
+      interactionFormService = proxyquire('../../src/services/interaction-form.service', {
         '../repositorys/interactionrepository': {
           saveInteraction: saveInteractionStub
         }
