+++ test/services/service-delivery.service.test.js
@@ -1,6 +1,6 @@
 /* globals expect: true, describe: true, it: true */
 /* eslint no-unused-expressions: 0 */
-const servicedeliveryservice = require('../../src/services/servicedeliveryservice')
+const servicedeliveryservice = require('../../src/services/service-delivery.service')
 
 describe('Service delivery formatting service', function () {
   describe('convertServiceDeliveryFormToApiFormat', function () {
